requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
cryptography==41.0.4
sqlalchemy==2.0.22
alembic==1.12.0
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
